- Moved Allow-Duplicated processing to RPMPackageData. This makes it be   calculated only once. - Now caching duplicated packages in RPMPackageData, so that it's not   necessary to execute regexes in packages we already know are duplicated.   Also, it is possible to set duplicated packages with the SetDupPackage(). - Cache the package name in rpmListParser, so that it is calculated only once   for each package. - Minor optimizations in version string building. - Implemented dynamic virtualization of packages! Here is the mail sent to   apt-rpm list:
I'd like to share with you my happiness about an interesting improvement
made in the in-development APT-RPM. Now Allow-Duplicated will also work
automatically for packages with multiple installed versions, and not
already in the Allow-Duplicated option. It means that there's no
database situation that APT-RPM or Synaptic won't be able to handle.

Notice that the Allow-Duplicated option won't go away, since for
packages like kernel, it's interesting to have that setting
hardcoded, even if only one package is installed in the system.

Also, packages in this situation won't be cleanly updated if new
versions are found in the server (just like those hardcoded in
Allow-Duplicated), so a big warning with instructions on how to
fix the situation is still issued to warn the user about what's
happening. It is possible, however, to disable that warning
including the option

RPM::Allow-Duplicated-Warning "false";

in the configuration file. Everything will work fine with this option,
but, as I said above, packages won't be cleanly updated if multiple
versions are found, so it might be a good idea to leave the warning
enabled, and include in the Allow-Duplicated option packages which
should stay in this situation (as the warning will tell the user, if
enabled).